NAV-29334: Forenkler komponenter i routes
