+++ src/frontend/sider/Fagsak/Dokumenter/Dokumenter.tsx
@@ -1,0 +1,11 @@
+import { Fagsaklinje } from '@komponenter/Saklinje/Fagsaklinje';
+import { JournalpostListe } from '@sider/Fagsak/Dokumenter/JournalpostListe';
+
+export function Dokumenter() {
+    return (
+        <>
+            <Fagsaklinje />
+            <JournalpostListe />
+        </>
+    );
+}
