+++ src/frontend/sider/Fagsak/Dokumenter/journalpostUtils.ts
