+++ src/frontend/sider/Fagsak/Dokumentutsending/Dokumentutsending.tsx
@@ -1,3 +1,5 @@
+import { Fagsaklinje } from '@komponenter/Saklinje/Fagsaklinje';
+import { fagsakHeaderHøydeRem } from '@typer/styling';
 import { useNavigate } from 'react-router';
 import styled from 'styled-components';
 
@@ -6,7 +8,6 @@ import { RessursStatus } from '@navikt/familie-typer';
 
 import { useDokumentutsendingContext } from './DokumentutsendingContext';
 import { DokumentutsendingSkjema } from './DokumentutsendingSkjema';
-import { fagsakHeaderHøydeRem } from '../../../typer/styling';
 import { useFagsakContext } from '../FagsakContext';
 
 const Container = styled.div`
@@ -23,45 +24,47 @@ export function Dokumentutsending() {
     const { hentetDokument, settVisInnsendtBrevModal, visInnsendtBrevModal } = useDokumentutsendingContext();
 
     return (
-        <Container>
-            {visInnsendtBrevModal && (
-                <Modal
-                    open
-                    onClose={() => settVisInnsendtBrevModal(false)}
-                    header={{ heading: 'Brevet er sendt', size: 'medium' }}
-                    portal
-                >
-                    <Modal.Footer>
-                        <Button
-                            variant={'secondary'}
-                            key={'til saksoversikt'}
-                            size={'medium'}
-                            onClick={() => {
-                                navigate(`/fagsak/${fagsak.id}/saksoversikt`);
-                                settVisInnsendtBrevModal(false);
-                            }}
-                            children={'Se saksoversikt'}
-                        />
-                        <Button
-                            variant={'secondary'}
-                            key={'til oppgavebenken'}
-                            size={'medium'}
-                            onClick={() => {
-                                navigate('/oppgaver');
-                            }}
-                            children={'Se oppgavebenk'}
-                        />
-                    </Modal.Footer>
-                </Modal>
-            )}
-            <DokumentutsendingSkjema />
-
-            <iframe
-                title={'dokument'}
-                src={hentetDokument.status === RessursStatus.SUKSESS ? hentetDokument.data : ''}
-                width={'100%'}
-                height={'100%'}
-            />
-        </Container>
+        <>
+            <Fagsaklinje />
+            <Container>
+                {visInnsendtBrevModal && (
+                    <Modal
+                        open
+                        onClose={() => settVisInnsendtBrevModal(false)}
+                        header={{ heading: 'Brevet er sendt', size: 'medium' }}
+                        portal
+                    >
+                        <Modal.Footer>
+                            <Button
+                                variant={'secondary'}
+                                key={'til saksoversikt'}
+                                size={'medium'}
+                                onClick={() => {
+                                    navigate(`/fagsak/${fagsak.id}/saksoversikt`);
+                                    settVisInnsendtBrevModal(false);
+                                }}
+                                children={'Se saksoversikt'}
+                            />
+                            <Button
+                                variant={'secondary'}
+                                key={'til oppgavebenken'}
+                                size={'medium'}
+                                onClick={() => {
+                                    navigate('/oppgaver');
+                                }}
+                                children={'Se oppgavebenk'}
+                            />
+                        </Modal.Footer>
+                    </Modal>
+                )}
+                <DokumentutsendingSkjema />
+                <iframe
+                    title={'dokument'}
+                    src={hentetDokument.status === RessursStatus.SUKSESS ? hentetDokument.data : ''}
+                    width={'100%'}
+                    height={'100%'}
+                />
+            </Container>
+        </>
     );
 }
