+++ src/frontend/sider/Fagsak/FagsakRoutes.tsx
@@ -1,13 +1,12 @@
 import { NotFound } from '@komponenter/Error/NotFound';
 import { RouteError } from '@komponenter/Error/RouteError';
-import { Fagsaklinje } from '@komponenter/Saklinje/Fagsaklinje';
-import { BehandlingContainer } from '@sider/Fagsak/Behandling/BehandlingContainer';
+import { Behandling } from '@sider/Fagsak/Behandling/Behandling';
 import { behandlingRoutes } from '@sider/Fagsak/Behandling/BehandlingRoutes';
 import { HentOgSettBehandlingProvider } from '@sider/Fagsak/Behandling/context/HentOgSettBehandlingContext';
+import { Dokumenter } from '@sider/Fagsak/Dokumenter/Dokumenter';
 import { Dokumentutsending } from '@sider/Fagsak/Dokumentutsending/Dokumentutsending';
 import { DokumentutsendingProvider } from '@sider/Fagsak/Dokumentutsending/DokumentutsendingContext';
-import { InfotrygdFagsak } from '@sider/Fagsak/Infotrygd/InfotrygdFagsak';
-import { JournalpostListe } from '@sider/Fagsak/journalposter/JournalpostListe';
+import { Infotrygd } from '@sider/Fagsak/Infotrygd/Infotrygd';
 import { RedirectTilSaksoversikt } from '@sider/Fagsak/Saksoversikt/RedirectTilSaksoversikt';
 import { Saksoversikt } from '@sider/Fagsak/Saksoversikt/Saksoversikt';
 import type { RouteObject } from 'react-router';
@@ -19,47 +18,29 @@ export const fagsakRoutes: RouteObject[] = [
     },
     {
         path: 'saksoversikt',
-        element: (
-            <>
-                <Fagsaklinje />
-                <Saksoversikt />
-            </>
-        ),
+        element: <Saksoversikt />,
     },
     {
         path: 'dokumentutsending',
         element: (
-            <>
-                <Fagsaklinje />
-                <DokumentutsendingProvider>
-                    <Dokumentutsending />
-                </DokumentutsendingProvider>
-            </>
+            <DokumentutsendingProvider>
+                <Dokumentutsending />
+            </DokumentutsendingProvider>
         ),
     },
     {
         path: 'dokumenter',
-        element: (
-            <>
-                <Fagsaklinje />
-                <JournalpostListe />
-            </>
-        ),
+        element: <Dokumenter />,
     },
     {
         path: 'infotrygd',
-        element: (
-            <>
-                <Fagsaklinje />
-                <InfotrygdFagsak />
-            </>
-        ),
+        element: <Infotrygd />,
     },
     {
         path: ':behandlingId',
         element: (
             <HentOgSettBehandlingProvider>
-                <BehandlingContainer />
+                <Behandling />
             </HentOgSettBehandlingProvider>
         ),
         errorElement: <RouteError />,
